# meta/openapi.yaml
openapi: 3.0.1
info:
  title: Sample API
  description: API description in Markdown.
  version: 1.0.0

servers:
  - description: Production GCloud
    url: "http://beta1.tracker.infra.teleport-park.com"

paths:
  /events:
    get:
      summary: Shows latest events
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Event" }

    post:
      summary: Tracks selected status
      description: Devices sends tracking info into this endpoint
      requestBody: { $ref: "#/components/requestBodies/EventRequest" }
      responses:
        201:
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Event" }

  /runs:
    get:
      summary: Returns list of runs
      parameters:
        - in: query
          name: o
          description: Offset for search
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: l
          description: Limit for search
          schema: { type: integer, minimum: 0, default: 100 }

      responses:
        200:
          description: OK
          headers:
            Content-Range:
              description: Content range for pagination
              schema: { type: string , example: 'items 0-10/100' }
            X-Total-Items:
              description: Total items in collection
              schema: { type: integer }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Run"}
            text/csv:
              schema:
                type: string

components:

  schemas:

    MachineType:
      type: string
      enum: [ tvr, tng, tpg ]
      description: "Machine type"

    MachineId:
      type: string
      format: uuid
      description: "Machine unique ID"

    Event:
      type: object
      required: [ id, type, status, timestamp ]
      properties:
        id: { type: string, description: "Machine ID or token" }
        type: { type: string, enum: [ tvr, tng, tpg ], description: "Machine type" }
        status: { type: string, enum: [ idle, playing ]}
        comment: { type: string, description: "Comment to know" }
        timestamp: { type: string, format: 'date-time' }
        local_timestamp: { type: string, format: 'date-time' }
        game: { $ref: "#/components/schemas/EventGame" }

    EventGame:
      type: object
      required: [ id ]
      properties:
        id: { type: string, description: "Game ID or UUID" }
        name: { type: string, description: "Game name" }

    Run:
      type: object
      required: [ id, type, timestamp ]
      properties:
        id: { $ref: "#/components/schemas/MachineId" }
        type: { $ref: "#/components/schemas/MachineType" }
        location: { type: string, description: "Machine location", example: "Sochi location 1" }
        start_at:  { type: string, format: 'date-time', description: "Game run start timestamp" }
        end_at:  { type: string, format: 'date-time', description: "Game run end timestamp" }
#        game: { $ref: "#/" }

  requestBodies:
    EventRequest:
      required: true
      description: Tracking info
      content:
        application/json:
          schema:
            type: object
            required: [ id, type, status ]
            properties:
              id: { $ref: "#/components/schemas/Event/properties/id" }
              type: { $ref: "#/components/schemas/Event/properties/type" }
              status: { $ref: "#/components/schemas/Event/properties/status" }
              comment: { $ref: "#/components/schemas/Event/properties/comment" }
              local_timestamp: { $ref: "#/components/schemas/Event/properties/local_timestamp" }
              game: { $ref: "#/components/schemas/Event/properties/game" }